Refactor(core-android): Add change listener to handle the secure display setting
- setRecentsScreenshotEnabled is always off even if secure display is turned off
